drm/compositor: Refactor exporter out
