+++ src/backend/drm/compositor/mod.rs
@@ -120,14 +120,13 @@
 //! }
 //! ```
 use std::{
-    cell::RefCell,
     collections::{HashMap, HashSet},
     fmt::Debug,
     io::ErrorKind,
     os::unix::io::{AsFd, OwnedFd},
     rc::Rc,
     str::FromStr,
-    sync::{Arc, Mutex},
+    sync::Arc,
 };
 
 use drm::{
@@ -173,13 +172,13 @@ use crate::{
 };
 
 use super::{
-    error::AccessError, surface::VrrSupport, DrmDeviceFd, DrmSurface, Framebuffer, PlaneClaim, PlaneInfo,
-    Planes,
+    error::AccessError,
+    exporter::{ExportBuffer, ExportFramebuffer},
+    surface::VrrSupport,
+    DrmSurface, Framebuffer, PlaneClaim, PlaneInfo, Planes,
 };
 
-pub mod dumb;
 mod elements;
-pub mod gbm;
 
 use elements::*;
 
@@ -841,98 +840,6 @@ impl<B: Framebuffer> FrameState<B> {
                         .and_then(|(_, fence)| fence.as_ref().map(|fence| fence.as_fd())),
                 }),
             })
-    }
-}
-
-/// Possible buffers to export as a framebuffer using [`ExportFramebuffer`]
-#[derive(Debug)]
-pub enum ExportBuffer<'a, B: Buffer> {
-    /// A wayland buffer
-    Wayland(&'a WlBuffer),
-    /// A [`Allocator`] buffer
-    Allocator(&'a B),
-}
-
-impl<'a, B: Buffer> ExportBuffer<'a, B> {
-    #[inline]
-    fn from_underlying_storage(storage: &'a UnderlyingStorage<'_>) -> Option<Self> {
-        match storage {
-            UnderlyingStorage::Wayland(buffer) => Some(Self::Wayland(buffer)),
-            UnderlyingStorage::Memory { .. } => None,
-        }
-    }
-}
-
-/// Export a [`ExportBuffer`] as a framebuffer
-pub trait ExportFramebuffer<B: Buffer>
-where
-    B: Buffer,
-{
-    /// Type of the framebuffer
-    type Framebuffer: Framebuffer;
-
-    /// Type of the error
-    type Error: std::error::Error;
-
-    /// Add a framebuffer for the specified buffer
-    fn add_framebuffer(
-        &self,
-        drm: &DrmDeviceFd,
-        buffer: ExportBuffer<'_, B>,
-        use_opaque: bool,
-    ) -> Result<Option<Self::Framebuffer>, Self::Error>;
-
-    /// Test if the provided buffer is eligible for adding a framebuffer
-    fn can_add_framebuffer(&self, buffer: &ExportBuffer<'_, B>) -> bool;
-}
-
-impl<F, B> ExportFramebuffer<B> for Arc<Mutex<F>>
-where
-    F: ExportFramebuffer<B>,
-    B: Buffer,
-{
-    type Framebuffer = <F as ExportFramebuffer<B>>::Framebuffer;
-    type Error = <F as ExportFramebuffer<B>>::Error;
-
-    #[inline]
-    fn add_framebuffer(
-        &self,
-        drm: &DrmDeviceFd,
-        buffer: ExportBuffer<'_, B>,
-        use_opaque: bool,
-    ) -> Result<Option<Self::Framebuffer>, Self::Error> {
-        let guard = self.lock().unwrap();
-        guard.add_framebuffer(drm, buffer, use_opaque)
-    }
-
-    #[inline]
-    fn can_add_framebuffer(&self, buffer: &ExportBuffer<'_, B>) -> bool {
-        let guard = self.lock().unwrap();
-        guard.can_add_framebuffer(buffer)
-    }
-}
-
-impl<F, B> ExportFramebuffer<B> for Rc<RefCell<F>>
-where
-    F: ExportFramebuffer<B>,
-    B: Buffer,
-{
-    type Framebuffer = <F as ExportFramebuffer<B>>::Framebuffer;
-    type Error = <F as ExportFramebuffer<B>>::Error;
-
-    #[inline]
-    fn add_framebuffer(
-        &self,
-        drm: &DrmDeviceFd,
-        buffer: ExportBuffer<'_, B>,
-        use_opaque: bool,
-    ) -> Result<Option<Self::Framebuffer>, Self::Error> {
-        self.borrow().add_framebuffer(drm, buffer, use_opaque)
-    }
-
-    #[inline]
-    fn can_add_framebuffer(&self, buffer: &ExportBuffer<'_, B>) -> bool {
-        self.borrow().can_add_framebuffer(buffer)
     }
 }
