+++ src/backend/drm/mod.rs
@@ -76,8 +76,10 @@ pub(crate) mod device;
 #[cfg(feature = "backend_drm")]
 pub mod dumb;
 mod error;
+pub mod exporter;
 #[cfg(feature = "backend_gbm")]
 pub mod gbm;
+pub mod output;
 
 mod surface;
